requirements.txt
cffi==1.9.1
cryptography==1.7.1
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.28.0
django-cors-headers==1.3.1
django-filter==0.15.3
django-rest-auth==0.8.2
djangorestframework==3.5.3
idna==2.1
ndg-httpsclient==0.4.2
oauthlib==2.0.0
Pillow==3.4.2
psycopg2==2.6.2
pusher==1.6
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
urllib3==1.17
